+++ earth2studio/data/gfs.py
@@ -323,7 +323,8 @@ class GFS:
             task.gfs_byte_length,
             task.gfs_modifier,
         )
-        xr_array[task.data_array_indices] = out
+        i, j, k = task.data_array_indices
+        xr_array[i, j, k] = out
 
     async def fetch_array(
         self,