+++ earth2studio/data/hrrr.py
@@ -404,7 +404,8 @@ class HRRR:
             task.hrrr_byte_length,
             task.hrrr_modifier,
         )
-        xr_array[task.data_array_indices] = out
+        i, j, k = task.data_array_indices
+        xr_array[i, j, k] = out
 
     async def fetch_array(
         self,