+++ test/data/test_wb2era5.py
@@ -17,6 +17,7 @@
 import datetime
 import pathlib
 import shutil
+from importlib.metadata import version
 
 import numpy as np
 import pytest
@@ -26,6 +27,9 @@ from earth2studio.data import WB2ERA5, WB2ERA5_32x64, WB2ERA5_121x240
 
 @pytest.mark.slow
 @pytest.mark.xfail
+@pytest.mark.skipif(
+    int(version("zarr").split(".")[0]) < 3, reason="Test requires zarr version > 3.0"
+)
 @pytest.mark.timeout(15)
 @pytest.mark.parametrize(
     "time",
@@ -62,6 +66,9 @@ def test_wb2era5_fetch(time, variable, Datasource):
 
 @pytest.mark.slow
 @pytest.mark.xfail
+@pytest.mark.skipif(
+    int(version("zarr").split(".")[0]) < 3, reason="Test requires zarr version > 3.0"
+)
 @pytest.mark.timeout(15)
 @pytest.mark.parametrize(
     "time",
@@ -102,6 +109,9 @@ def test_wb2era5_cache(time, variable, cache, Datasource):
 
 @pytest.mark.xfail
 @pytest.mark.timeout(15)
+@pytest.mark.skipif(
+    int(version("zarr").split(".")[0]) < 3, reason="Test requires zarr version > 3.0"
+)
 @pytest.mark.parametrize(
     "time",
     [